ci(dep): move out from main.yml
